+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Format follows [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+- add `window.opacity` for a transparent background (compositor-dependent)
+
 ## [0.10.0] - 2026-07-25
 
 ### Added